requirements.txt
# Core dependencies
requests>=2.28.0
rapidfuzz>=3.0.0
python-dotenv>=1.0.0

# Optional enhancements (uncomment as needed)
# pandas>=2.0.0        # For advanced data analysis
# matplotlib>=3.7.0    # For visualization
# schedule>=1.2.0      # For automated scheduling
# redis>=4.5.0         # For caching
